fix: move logic to store
