+++ packages/builder/src/pages/builder/workspace/[application]/agent/[agentId]/deployment.svelte
@@ -8,17 +8,8 @@
     notifications,
   } from "@budibase/bbui"
   import { FeatureFlag, type Agent, type DeploymentRow } from "@budibase/types"
-  import {
-    selectedAgent,
-    agentsStore,
-    chatAppsStore,
-    currentChatApp,
-    featureFlags,
-  } from "@/stores/portal"
-  import { deploymentStore } from "@/stores/builder"
-  import { params } from "@roxi/routify"
+  import { selectedAgent, agentsStore, featureFlags } from "@/stores/portal"
   import AgentChatChannel from "./DeploymentChannels/AgentChatChannel.svelte"
-  import { buildAgentChatUpdate } from "./DeploymentChannels/agentChatDeployment"
   import DiscordConfig from "./DeploymentChannels/DiscordConfig.svelte"
   import DiscordLogo from "assets/discord.svg"
 
@@ -28,8 +19,6 @@
   let currentAgent: Agent | undefined = $derived($selectedAgent)
   let discordModal: Modal
   let toggling = $state(false)
-  let togglingChat = $state(false)
-  let loadingChatApp = $state(false)
 
   const discordConfigured = $derived.by(() => {
     const integration = currentAgent?.discordIntegration
@@ -47,39 +36,6 @@
 
   const hasAiConfig = $derived.by(() => !!currentAgent?.aiconfig?.trim())
   const agentChatEnabled = $derived(!!$featureFlags[FeatureFlag.AI_CHAT])
-  const chatAgentEnabled = $derived.by(() => {
-    const agentId = currentAgent?._id
-    if (!agentId) {
-      return false
-    }
-
-    const chatAgent = ($currentChatApp?.agents || []).find(
-      agent => agent.agentId === agentId
-    )
-    return !!chatAgent?.isEnabled
-  })
-
-  $effect(() => {
-    const workspaceId = $params.application
-    if (
-      !agentChatEnabled ||
-      !workspaceId ||
-      !!$currentChatApp ||
-      loadingChatApp
-    ) {
-      return
-    }
-
-    loadingChatApp = true
-    void chatAppsStore
-      .ensureChatApp(undefined, workspaceId)
-      .catch(error => {
-        console.error(error)
-      })
-      .finally(() => {
-        loadingChatApp = false
-      })
-  })
 
   const channels = $derived.by<DeploymentRow[]>(() => [
     {
@@ -129,53 +85,6 @@
       toggling = false
     }
   }
-
-  const toggleAgentChatDeployment = async () => {
-    const workspaceId = $params.application
-    const agentId = currentAgent?._id
-    if (!workspaceId || !agentId || togglingChat) {
-      return
-    }
-
-    togglingChat = true
-    let isCurrentlyEnabled = false
-
-    try {
-      const ensuredChatApp = await chatAppsStore.ensureChatApp(
-        undefined,
-        workspaceId
-      )
-      if (!ensuredChatApp) {
-        notifications.error("Could not update chat")
-        return
-      }
-
-      const chatAgents = ensuredChatApp.agents || []
-      isCurrentlyEnabled = chatAgents.some(
-        agent => agent.agentId === agentId && agent.isEnabled
-      )
-      const update = buildAgentChatUpdate({
-        agents: chatAgents,
-        agentId,
-        enable: !isCurrentlyEnabled,
-      })
-
-      await chatAppsStore.updateChatApp(update)
-      await deploymentStore.publishApp()
-      notifications.success(
-        isCurrentlyEnabled ? "Agent chat disabled" : "Agent chat enabled"
-      )
-    } catch (error) {
-      console.error(error)
-      notifications.error(
-        isCurrentlyEnabled
-          ? "Failed to disable agent chat"
-          : "Failed to enable agent chat"
-      )
-    } finally {
-      togglingChat = false
-    }
-  }
 </script>
 
 <div class="deployment-root">
@@ -211,12 +120,7 @@
     </div>
     <div class="integration-list">
       {#if agentChatEnabled}
-        <AgentChatChannel
-          agentId={currentAgent?._id}
-          enabled={chatAgentEnabled}
-          disabled={togglingChat || loadingChatApp || !currentAgent?._id}
-          onToggle={toggleAgentChatDeployment}
-        />
+        <AgentChatChannel agentId={currentAgent?._id} />
       {/if}
       {#each channels as channel (channel.id)}
         <div class="integration-row">
